backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.0.1
numpy==1.26.3
google-generativeai==0.3.2
python-dotenv==1.0.0
werkzeug==2.3.7
protobuf==3.20.3